created component folder for DB and removed code from App.js
